jf-corrections-entites

+++ src/main/java/com/example/demo/entities/Dish.java
@@ -5,7 +5,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.Date;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -17,7 +17,7 @@ public class Dish {
     @GeneratedValue
     private long id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private String description;
 
     @JsonIgnore
